user_service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.2.4
django-axes==7.0.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-ipware==7.0.1
django-otp==1.6.1
django-phonenumber-field==8.0.0
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dnspython==2.7.0
docopt==0.6.2
drf-spectacular==0.28.0
filelock==3.18.0
flower==2.0.1
frozenlist==1.7.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.4
Markdown==3.8.2
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.4.2
oauthlib==3.2.2
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.2.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.23.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
yarg==0.1.10
yarl==1.20.1
